perf(core): SIMD column-sum reductions in data_probe, drop redundant cast
- data_probe: route the four column-sum reductions in classify_all_sinks,
  classify_all_sinks_flat, and apply_dual_policy_gate_flat through
  simd_add_inplace (NEON/AVX2). Each was previously a scalar
  'for j in 0..m { col_sums[j] += row[j]; }' inner loop.
- sense/backend: drop the redundant 'as u32' intermediate when casting
  bit-tested u64 down to f32 ('as f32' works directly on u64).

+++ crates/katgpt-core/src/sense/backend.rs
@@ -219,8 +219,9 @@ fn project_ternary(
     for i in 0..n_directions {
         let dir = &directions[i];
         let hla_val = hla_state[i];
-        let pos = ((dir.pos_bits >> i) & 1) as u32 as f32;
-        let neg = ((dir.neg_bits >> i) & 1) as u32 as f32;
+        // Drop the redundant `as u32` intermediate — `as f32` works directly on `u64`.
+        let pos = ((dir.pos_bits >> i) & 1) as f32;
+        let neg = ((dir.neg_bits >> i) & 1) as f32;
         dot += (pos - neg) * hla_val * dir.row_scale;
     }
     confidence * crate::simd::fast_sigmoid(dot)
